feat: implemented Canvas and Geometry abstractions
The Canvas represents a windows and is in charge of managing the Arcball
transformation, resize of the windows, scroll for scale and button clicks.
The Canvas also maintains a list of geometries that are drawn during the
rendering.

The Geometry is an abstraction and is in charge of implementing a draw()
function that setup and draw the relative geometry using OpenGL.